requirements.txt
#Site Engine
Django==4.0.2

#Database
psycopg2==2.9.5
django-taggit==3.0.0

#Forms
crispy-bootstrap5==0.6

# Environment
python-dotenv==0.21.0